Update MilitaryOverlay-2525C.lpkx
I noticed while poking around getting ready to do some documentation that
- Stability group layer has 2525D in tags
- METOC points has 2525D in tags
- Credits "Esri" should be "Esri, Inc."

I ran "Delete All Features from Workspace" and "SpatialIndexWorkAround" on the entire .gdb before creating the schema only .lpkx. I did not update any source files as I was not sure which ones power the group layer descriptions, and in the case of METOC points it seems that the source metadata file that should be getting used does not match information present in the .lpkx (both are accurate, just slightly different).

This is not a highly critical change, so if it is deemed too high of risk, we can hold off, or just ignore