updates to .csv for app6B
the .gdb needs to be rebuilt to pick up required schema changes (that were in before this commit) for Land Equipment and Control Measures. Since we need to rebuild the .gdb anyway for these schema changes, I want to put in some other schema changes that are not explicitly required by the standard, but may be used depending on a user's/integrator's intrepretation of the standard.

@csmoore